src/web/app.py
"""
ZR Daily Report Web应用
基于FastAPI的Web管理界面
"""

import asyncio
import json
import os
import uuid
from datetime import datetime
from typing import Any, Dict, List, Optional

import uvicorn
from fastapi import BackgroundTasks, FastAPI, File, HTTPException, Request, UploadFile
from fastapi.responses import FileResponse, HTMLResponse
from fastapi.staticfiles import StaticFiles
from fastapi.templating import Jinja2Templates
from pydantic import BaseModel

# 创建FastAPI应用
app = FastAPI(
    title="ZR Daily Report Web",
    description="ZR Daily Report Web管理界面",
    version="1.0.0",
)


# 数据模型
class DeviceInfo(BaseModel):
    device_code: str
    start_date: str
    end_date: str


class ReportRequest(BaseModel):
    devices: List[DeviceInfo]
    report_type: str
    output_format: str = "xlsx"


@app.get("/")
async def index():
    """主页"""
    return {"message": "ZR Daily Report Web API"}


@app.get("/api/health")
async def health_check():
    """健康检查"""
    return {
        "status": "healthy",
        "timestamp": datetime.now().isoformat(),
        "version": "1.0.0",
    }


@app.post("/api/reports/generate")
async def generate_report(request: ReportRequest):
    """生成报表"""
    try:
        task_id = str(uuid.uuid4())
        return {
            "task_id": task_id,
            "status": "submitted",
            "message": "报表生成任务已提交",
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
