requirements.txt
# 基础依赖
openpyxl==3.1.2
mysql-connector-python==8.0.33
pandas>=1.5.0
cryptography==43.0.1

# Web框架和API
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# 异步支持
aiohttp>=3.8.0
aiomysql==0.2.0
asyncio-mqtt==0.16.1

# 缓存支持
redis==5.0.1
aioredis==2.0.1

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 测试依赖
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.21.1
httpx==0.25.2
mock>=4.0.0
tox>=3.25.0

# 开发工具
flake8==7.1.1
black==24.8.0
mypy>=0.971
isort==5.12.0
pre-commit>=3.3.0

# 构建工具
build>=0.10.0
twine>=4.0.0

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8

# 使用阿里云镜像源安装依赖:
# pip install -r requirements.txt -i https://mirrors.aliyun.com/pypi/simple/